support dmg and cgb bootrom, added bg fifo metadata (palette/prio), fixed tile attribute in fetcher

+++ include/renderer.h
@@ -9,7 +9,7 @@ public:
 	static constexpr int WIDTH = 160;
 	static constexpr int HEIGHT = 144;
 	bool init(const char* title, int scale);
-	void render(const std::array<uint8_t, WIDTH* HEIGHT>& framebuffer);
+	void render(const std::array<uint32_t, WIDTH* HEIGHT>& framebuffer);
 	bool procEvents();
 	void idle();
 	SDL_Window* getWindow() { return sdlWindow; }
@@ -21,5 +21,5 @@ private:
 	SDL_Texture* sdlTexture = nullptr;
 public:
 	std::array<uint32_t, WIDTH * HEIGHT> rgbbuffer;
-	std::array<uint32_t, WIDTH* HEIGHT> prevFrame{};
+	std::array<uint32_t, WIDTH * HEIGHT> prevFrame{};
 };
